feat: Add support for pgvecto-rs
