+++ neosearch/pyproject.toml
@@ -16,6 +16,8 @@ llama-index-llms-ollama = "^0.1.2"
 pyyaml = "^6.0.1"
 llama-index-embeddings-huggingface = "^0.1.4"
 llama-index-llms-llama-cpp = "^0.1.3"
+llama-index-vector-stores-pgvecto-rs = "^0.1.2"
+pgvecto-rs = {extras = ["sdk"], version = "^0.1.4"}
 
 [tool.poetry.dependencies.uvicorn]
 extras = [ "standard" ]
